+++ src/frontends/onnx/frontend/src/op/conv_transpose.cpp
@@ -17,7 +17,7 @@
 #include "openvino/op/strided_slice.hpp"
 #include "openvino/op/subtract.hpp"
 #include "utils/convpool.hpp"
-#include "utils/ov_builders/reshape.hpp"
+#include "utils/reshape.hpp"
 
 using namespace ov::op;
 using ov::CoordinateDiff;
