+++ src/frontends/onnx/frontend/src/op/gemm.cpp
@@ -8,7 +8,7 @@
 #include "openvino/op/constant.hpp"
 #include "openvino/op/matmul.hpp"
 #include "openvino/op/multiply.hpp"
-#include "utils/ov_builders/reshape.hpp"
+#include "utils/reshape.hpp"
 
 using namespace ov::op;
 using ov::Shape;
