+++ src/frontends/onnx/frontend/src/op/hardmax.cpp
@@ -13,7 +13,6 @@
 #include "openvino/op/shape_of.hpp"
 #include "openvino/op/topk.hpp"
 #include "utils/common.hpp"
-#include "utils/ov_builders/reshape.hpp"
 #include "utils/reshape.hpp"
 #include "validation_util.hpp"
