+++ src/frontends/onnx/frontend/src/op/log_softmax.cpp
@@ -10,7 +10,7 @@
 #include "openvino/op/log_softmax.hpp"
 #include "openvino/op/reshape.hpp"
 #include "openvino/op/shape_of.hpp"
-#include "utils/ov_builders/reshape.hpp"
+#include "utils/reshape.hpp"
 #include "validation_util.hpp"
 
 using namespace ov::op;
